Refactor chat_with_tj recipe: update README, add index.js, and streamline configuration
