+++ recipes/chat_with_tj/ibm-credentials.sample.env
@@ -1,9 +1,2 @@
-# This is a sample of how IBM credentials are stored. Values after the equals sign have been removed.
-# Please download your credentials for the services below and place them in a file named
-# ibm-credentials.env in this directory.
-SPEECH_TO_TEXT_APIKEY=
-SPEECH_TO_TEXT_URL=
-TEXT_TO_SPEECH_APIKEY=
-TEXT_TO_SPEECH_URL=
 WATSONX_AI_AUTH_TYPE=iam
-WATSONX_AI_APIKEY=
+WATSONX_AI_APIKEY= # FILL IN WITH YOUR IBM CLOUD API KEY
